more tidy

+++ test/communicator_all_gatherv.cpp
@@ -4,7 +4,8 @@
 #include "../../mpi3/main.hpp"
 #include "../../mpi3/communicator.hpp"
 #include "../../mpi3/detail/iterator.hpp"
-#include<numeric> //iota
+
+#include<numeric>  // for std::iota
 
 namespace mpi3 = boost::mpi3;
 
@@ -15,20 +16,20 @@ auto mpi3::main(int/*argc*/, char**/*argv*/, mpi3::communicator world) -> int tr
 {
 	using T = std::tuple<double, double>;
 	std::vector<T> v_local(10, T{world.rank(), world.rank()});
-	std::vector<T> v(v_local.size()*world.size());
+	std::vector<T> v( v_local.size() * static_cast<std::size_t>(world.size()) );
 	auto end = world.all_gather_n(v_local.begin(), v_local.size(), v.begin());
 	assert(end == v.end());
-	assert(( v[ 0] == T{0.,0.} ));
-	assert(( v[ 9] == T{0.,0.} ));
-	assert(( v[10] == T{1.,1.} ));
-	assert(( v[19] == T{1.,1.} ));
-	assert(( v[20] == T{2.,2.} ));
-	assert(( v[29] == T{2.,2.} ));
+	assert(( v[ 0] == T{0.0, 0.0} ));
+	assert(( v[ 9] == T{0.0, 0.0} ));
+	assert(( v[10] == T{1.0, 1.0} ));
+	assert(( v[19] == T{1.0, 1.0} ));
+	assert(( v[20] == T{2.0, 2.0} ));
+	assert(( v[29] == T{2.0, 2.0} ));
 }
 {
 	using T = std::tuple<double, double>;
-	std::vector<T> v_local(world.rank() + 5, T{world.rank(), world.rank()});
-	std::vector<T> v(1000, T{-99., -99.});
+	std::vector<T> v_local(static_cast<std::size_t>(world.rank() + 5), T{world.rank(), world.rank()});
+	std::vector<T> v(1000, T{-99.0, -99.0});
 	auto d_last = world.all_gatherv_n(begin(v_local), v_local.size(), begin(v));
 
 	// int predict_size = 0;
@@ -46,7 +47,7 @@ auto mpi3::main(int/*argc*/, char**/*argv*/, mpi3::communicator world) -> int tr
 }
 {
 	using T = std::tuple<double, double>;
-	std::vector<T> v_local(world.rank() + 5, T{world.rank(), world.rank()});
+	std::vector<T> v_local(static_cast<std::size_t>(world.rank() + 5), T{world.rank(), world.rank()});
 	std::vector<T> v(1000, T{-99.0, -99.0});
 	auto d_last = world.all_gatherv(begin(v_local), end(v_local), begin(v));
 	assert( d_last < end(v) );
@@ -68,10 +69,8 @@ auto mpi3::main(int/*argc*/, char**/*argv*/, mpi3::communicator world) -> int tr
 //	v.shrink_to_fit(); // to save _some_ memory
 
 // check communication
-	assert((v==std::vector<T>{0., 0., 0., 1., 1., 1., 1., 2., 2., 2., 2., 2.}));
+	assert((v==std::vector<T>{0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 2.0, 2.0}));
 }
 	return 0;
-}catch(...){
-	return 0;
-}
+} catch(...) {return 0;}
